refactor: split constants into different files
