+++ dev/constants/physical.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+RELAXED_MUSCLE_LENGTH_ONE = 9
+RELAXED_MUSCLE_LENGTH_TWO = 12
